requirements.txt
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.14.5
click==6.7
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
greenlet==1.0.0
itsdangerous==0.24
Jinja2==2.10
joblib==1.0.1
MarkupSafe==1.1.1
mysqlclient==2.0.3
Pillow==8.1.2
pycparser==2.18
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.2.7
threadpoolctl==2.1.0
Werkzeug==0.14.1
WTForms==2.1
xgboost==1.3.3
urllib3==1.26.4
Werkzeug
numpy==1.19.5
pandas==1.1.5
gunicorn==19.9.0
